# gateway/swagger/swagger.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Microservices Gateway API",
    "description": "HTTP Gateway для микросервисной архитектуры",
    "version": "1.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/github.com.krus210.balun_microservices.protobuf.users.v1.proto.UsersService/CreateProfile": {
      "post": {
        "summary": "CreateProfile - Создание профиля пользователя",
        "operationId": "UsersService_CreateProfile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoCreateProfileResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoCreateProfileRequest"
            }
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    },
    "/github.com.krus210.balun_microservices.protobuf.users.v1.proto.UsersService/GetProfileByID": {
      "post": {
        "summary": "GetProfileByID - Получение профиля пользователя по ID",
        "operationId": "UsersService_GetProfileByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoGetProfileByIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoGetProfileByIDRequest"
            }
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    },
    "/github.com.krus210.balun_microservices.protobuf.users.v1.proto.UsersService/GetProfileByNickname": {
      "post": {
        "summary": "GetProfileByNickname - Получение профиля пользователя по никнейму",
        "operationId": "UsersService_GetProfileByNickname",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoGetProfileByNicknameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoGetProfileByNicknameRequest"
            }
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    },
    "/github.com.krus210.balun_microservices.protobuf.users.v1.proto.UsersService/SearchByNickname": {
      "post": {
        "summary": "SearchByNickname - Поиск профиля пользователя по никнейму",
        "operationId": "UsersService_SearchByNickname",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoSearchByNicknameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoSearchByNicknameRequest"
            }
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    },
    "/github.com.krus210.balun_microservices.protobuf.users.v1.proto.UsersService/UpdateProfile": {
      "post": {
        "summary": "UpdateProfile - Обновление профиля пользователя",
        "operationId": "UsersService_UpdateProfile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoUpdateProfileResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoUpdateProfileRequest"
            }
          }
        ],
        "tags": [
          "UsersService"
        ]
      }
    }
  },
  "definitions": {
...
    "protoCreateProfileRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64",
          "title": "userId - идентификатор пользователя"
        },
        "nickname": {
          "type": "string",
          "title": "nickname - никнейм пользователя"
        },
        "bio": {
          "type": "string",
          "title": "bio - биография пользователя"
        },
        "avatarUrl": {
          "type": "string",
          "title": "avatarUrl - ссылка на аватар пользователя"
        }
      },
      "title": "CreateProfileRequest - запрос CreateProfile"
    },
    "protoCreateProfileResponse": {
      "type": "object",
      "properties": {
        "userProfile": {
          "$ref": "#/definitions/protoUserProfile",
          "title": "userProfile - профиль пользователя"
        }
      },
      "title": "CreateProfileResponse - ответ CreateProfile"
    },
...
    "protoGetProfileByIDRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64",
          "title": "userId - идентификатор пользователя"
        }
      },
      "title": "GetProfileByIDRequest - запрос GetProfileByID"
    },
    "protoGetProfileByIDResponse": {
      "type": "object",
      "properties": {
        "userProfile": {
          "$ref": "#/definitions/protoUserProfile",
          "title": "userProfile - профиль пользователя"
        }
      },
      "title": "GetProfileByIDResponse - ответ GetProfileByID"
    },
    "protoGetProfileByNicknameRequest": {
      "type": "object",
      "properties": {
        "nickname": {
          "type": "string",
          "title": "nickname - никнейм пользователя"
        }
      },
      "title": "GetProfileByNicknameRequest - запрос GetProfileByNickname"
    },
    "protoGetProfileByNicknameResponse": {
      "type": "object",
      "properties": {
        "userProfile": {
          "$ref": "#/definitions/protoUserProfile",
          "title": "userProfile - профиль пользователя"
        }
      },
      "title": "GetProfileByNicknameResponse - ответ GetProfileByNickname"
    },
...
    "protoSearchByNicknameRequest": {
      "type": "object",
      "properties": {
        "query": {
          "type": "string",
          "title": "query - запрос для поиска"
        },
        "limit": {
          "type": "string",
          "format": "int64",
          "title": "limit - лимит результатов"
        }
      },
      "title": "SearchByNicknameRequest - запрос SearchByNickname"
    },
    "protoSearchByNicknameResponse": {
      "type": "object",
      "properties": {
        "results": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protoUserProfile"
          },
          "title": "results - список профилей пользователей"
        }
      },
      "title": "SearchByNicknameResponse - ответ SearchByNickname"
    },
...
    "protoUpdateProfileRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64",
          "title": "userId - идентификатор пользователя"
        },
        "nickname": {
          "type": "string",
          "title": "nickname - никнейм пользователя"
        },
        "bio": {
          "type": "string",
          "title": "bio - биография пользователя"
        },
        "avatarUrl": {
          "type": "string",
          "title": "avatarUrl - ссылка на аватар пользователя"
        }
      },
      "title": "UpdateProfileRequest - запрос UpdateProfile"
    },
    "protoUpdateProfileResponse": {
      "type": "object",
      "properties": {
        "userProfile": {
          "$ref": "#/definitions/protoUserProfile",
          "title": "userProfile - профиль пользователя"
        }
      },
      "title": "UpdateProfileResponse - ответ UpdateProfile"
    },
    "protoUserProfile": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64",
          "title": "userId - идентификатор пользователя"
        },
        "nickname": {
          "type": "string",
          "title": "nickname - никнейм пользователя"
        },
        "bio": {
          "type": "string",
          "title": "bio - биография пользователя"
        },
        "avatarUrl": {
          "type": "string",
          "title": "avatarUrl - ссылка на аватар пользователя"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}